nomad-3/11-pomodoro : Refactor to split widgets
